--- src/Pages/Detail/Slider/Silder.scss
@@ -1,12 +1,0 @@
-.anotherItemWrapper {
-  overflow: hidden;
-
-  .anotherItemList {
-    position: relative;
-    top: 0px;
-    left: 1.5px;
-    display: flex;
-    justify-content: flex-start;
-    transition: transform 0.5s ease-in-out;
-  }
-}
